refactor: update auth service integration in conversation controller
