+++ backend/src/modules/auth/auth.route.ts
@@ -1,17 +1,7 @@
 import { Router } from "express";
 
-import authService from "./auth.controller";
-import {
-  emailSchema,
-  forgotPasswordSchema,
-  regenerateCodeSchema,
-  userLoginSchema,
-  userRegisterSchema,
-  verifyEmailSchema,
-} from "./auth.schema";
 import { asyncHandler } from "../../utils/asyncHandler";
-import { validateZod } from "../../utils/validationHandler";
-const Controller = authService;
+import { authController } from "./auth.controller";
 
 const router = Router();
 
@@ -22,70 +12,17 @@ router.get("/", (req, res, next) => {
   });
 });
 
-router.post(
-  "/register",
-  asyncHandler(async (req, res) => {
-    const validationResult = validateZod(userRegisterSchema, req.body);
-    await Controller.create(validationResult, res);
-  })
-);
-
-router.post(
-  "/verify",
-  asyncHandler(async (req, res) => {
-    const validationResult = validateZod(verifyEmailSchema, req.body);
-
-    if (!validationResult.success) throw new Error("Invalid Request Body");
-    await Controller.verifyEmail(validationResult, res);
-  })
-);
-
-router.post(
-  "/regenerate",
-  asyncHandler(async (req, res) => {
-    const validationResult = validateZod(regenerateCodeSchema, req.body);
-    await Controller.regenerate(validationResult.email, res);
-  })
-);
-
-router.post(
-  "/verifyable-email",
-  asyncHandler(async (req, res) => {
-    const validEmail = emailSchema.parse(req.body.email);
-    await Controller.verifyAbleEmail(validEmail, res);
-  })
-);
-
+router.post("/register", asyncHandler(authController.register));
+router.post("/verify", asyncHandler(authController.verifyEmail));
+router.post("/regenerate", asyncHandler(authController.regenerate));
+router.post("/login", asyncHandler(authController.login));
+router.post("/verifyable-email", asyncHandler(authController.verifyAbleEmail));
 router.post(
   "/forgot-password-generator",
-  asyncHandler(async (req, res, next) => {
-    const validEmail = emailSchema.parse(req.body.email);
-    await Controller.forgotPasswordToken(validEmail, res);
-  })
+  asyncHandler(authController.forgotPasswordToken)
 );
-
-router.post(
-  "/forgot-password",
-  asyncHandler(async (req, res) => {
-    const validationResult = validateZod(forgotPasswordSchema, req.body);
-    await Controller.forgotPassword(validationResult, res);
-  })
-);
-
-router.post(
-  "/login",
-  asyncHandler(async (req, res) => {
-    const validationResult = validateZod(userLoginSchema, req.body);
-    await Controller.login(validationResult, res);
-  })
-);
-
-router.get("/me", asyncHandler(Controller.checkTokens));
-
-router.post("/logout", async (req, res, next) => {
-  res.clearCookie("access_token");
-  res.clearCookie("refresh_token");
-  return res.status(200).json({ message: "success" });
-});
+router.post("/forgot-password", asyncHandler(authController.forgotPassword));
+router.get("/me", asyncHandler(authController.me));
+router.post("/logout", authController.logout);
 
 export const authRouter = router;
